+++ resources/js/components/KanbanBoard.vue
@@ -2,37 +2,38 @@
     <div class="max-w-7xl flex-1 mx-auto flex flex-col overflow-auto sm:px-6 lg:px-8">
         <div class="w-full mb-6 flex">
             <Teleport to="body">
-                <generic-modal :show="kanban.creatingTask" @close="kanban.creatingTask = false" key="createTaskModal">
+                <generic-modal :show="kanban.creatingTask || kanban.updatingTask" @close="resetTasks()" key="createTaskModal">
                     <div>
                         <div class="mt-3 sm:mt-2">
-                            <DialogTitle as="h3" class="mb-6 text-base font-semibold leading-6 text-gray-900">Create a new task</DialogTitle>
+                            <DialogTitle as="h3" class="mb-6 text-base font-semibold leading-6 text-gray-900">{{ kanban.creatingTask ? 'Create a new task' : 'Update a task' }}</DialogTitle>
                             <div>
                                 <label for="name" class="block text-sm font-medium leading-6 text-gray-900">Task description</label>
                                 <div class="relative mt-2">
-                                    <input type="text" v-model="kanban.creatingTaskProps.name" id="name"
+                                    <input type="text" v-model="taskPropsName" id="name"
+                                        @input="validateForm()"
                                         class="peer block w-full border-0 bg-gray-50 py-1.5 text-gray-900 focus:ring-0 sm:text-sm sm:leading-6"
                                         placeholder="Make it productive, but also fun!" />
+                                    <div class="absolute inset-x-0 border-t border-gray-300 peer-focus:border-t-2 peer-focus:border-blue-600"
+                                        aria-hidden="true" />
                                     <p v-if="hasError('name')" 
-                                        class="mt-2 text-sm text-red-600" 
+                                        class="mt-2 ml-3 text-sm text-red-600"
                                         id="name-error">
                                             {{ getError('name') }}
                                     </p>
-                                    <div class="absolute inset-x-0 bottom-0 border-t border-gray-300 peer-focus:border-t-2 peer-focus:border-blue-600"
-                                        aria-hidden="true" />
                                 </div>
                             </div>
 
-                            <Listbox as="div" v-model="kanban.creatingTaskProps.user_id" class="mt-8">
+                            <Listbox as="div" v-model="taskPropsUser" class="mt-8">
                                 <ListboxLabel class="block text-sm font-medium leading-6 text-gray-900">Assigned to</ListboxLabel>
                                 <div class="relative mt-2">
                                     <ListboxButton
                                         class="relative w-full cursor-default rounded-md bg-white py-1.5 pl-3 pr-10 text-left text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:outline-none focus:ring-2 focus:ring-blue-500 sm:text-sm sm:leading-6">
                                         <span class="flex items-center">
-                                            <img :src="getAvatar(kanban.users[kanban.creatingTaskProps.user_id || kanban.self.id])" 
+                                            <img :src="getAvatar(kanban.users[taskPropsUser || kanban.self.id])"
                                                  alt="" 
                                                  class="h-5 w-5 flex-shrink-0 rounded-full" 
                                             />
-                                            <span class="ml-3 block truncate">{{ kanban.users[kanban.creatingTaskProps.user_id || kanban.self.id].name }}</span>
+                                            <span class="ml-3 block truncate">{{ kanban.users[taskPropsUser || kanban.self.id].name }}</span>
                                         </span>
                                         <span
                                             class="pointer-events-none absolute inset-y-0 right-0 ml-3 flex items-center pr-2">
@@ -66,7 +67,7 @@
 
                             <div class="mt-8">
                                 <label for="taskPhase" class="block text-sm font-medium leading-6 text-gray-900">Phase</label>
-                                <select v-model="kanban.creatingTaskProps.phase_id" id="taskPhase" class="mt-2 block w-full rounded-md border-0 py-1.5 pl-3 pr-10 text-gray-900 ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-blue-600 sm:text-sm sm:leading-6">
+                                <select v-model="taskPropsPhase" id="taskPhase" class="mt-2 block w-full rounded-md border-0 py-1.5 pl-3 pr-10 text-gray-900 ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-blue-600 sm:text-sm sm:leading-6">
                                     <option v-for="phase in kanban.phases" :key="phase.id" :value="phase.id">{{ phase.name }}</option>
                                 </select>
                             </div>
@@ -75,15 +76,19 @@
 
                         <div class="mt-5 sm:mt-6">
                             <button type="button"
+                                :disabled="kanban.submittingTask"
                                 class="inline-flex w-full justify-center rounded-md bg-blue-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-blue-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600"
-                                @click="addCard()">Add the card!</button>
+                                @click="addCard()">
+                                <span v-if="!kanban.submittingTask">{{ kanban.creatingTask ? 'Add the card!' : 'Update the card!' }}</span>
+                                <VueSpinnerTail v-else size="20" color="white" />
+                            </button>
                         </div>
                     </div>
                 </generic-modal>
             </Teleport>
         </div>
 
-        <div id="kanban-container" class="flex-1 flex overflow-auto scrollbar-hide shadow-lg">
+        <div id="kanban-container" ref="kanbanContainer" class="flex-1 flex overflow-auto scrollbar-hide shadow-lg">
             <div class="text-gray-900">
                 <div class="h-full flex overflow-x-auto overflow-y-auto space-x-4">
                     <task-column v-for="col in kanban.phases" :phase_id="col.id"></task-column>
@@ -103,9 +108,10 @@
                     <div class="mt-3 sm:mt-5">
                         <DialogTitle as="h3" class="text-base font-semibold leading-6 text-gray-900">{{ kanban.selectedTask.name }}</DialogTitle>
                         <div class="mt-2">
-                            <p class="text-sm text-gray-500">In column {{ kanban.phases[kanban.selectedTask.phase_id].name
+                            <p class="text-sm text-gray-500 mb-1">In column {{ kanban.phases[kanban.selectedTask.phase_id].name
                             }}</p>
-                            <p class="text-sm text-gray-500">Assigned to {{ kanban.selectedTask.user.name }}</p>
+                            <p class="text-sm text-gray-500 mb-1">Assigned to {{ kanban.selectedTask.user.name }}</p>
+                            <p v-if="kanban.selectedTask.completed_at" class="text-sm text-gray-500">Completed at {{ kanban.selectedTask.completed_at }}</p>
                         </div>
                     </div>
                 </div>
@@ -116,21 +122,76 @@
                 </div>
             </generic-modal>
         </Teleport>
+
+        <!-- Modal to confirm the actions -->
+        <Teleport to="body">
+            <generic-modal :show="kanban.openModal" @close="kanban.openModal = false" key="confirmActionModal">
+                <div>
+                    <div class="flex justify-center mt-3 sm:mt-2">
+                        <DialogTitle as="h3" class="mb-6 text-base font-semibold leading-6 text-gray-900">Are you sure?</DialogTitle>
+                    </div>
+
+                    <div class="flex justify-between mt-5 sm:mt-6">
+                        <button type="button"
+                            :disabled="kanban.confirmingAction"
+                            class="inline-flex w-full justify-center rounded-md mr-10 bg-blue-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-blue-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600"
+                            @click="confirmAction()">
+                            <span v-if="!kanban.confirmingAction">Yes</span>
+                            <VueSpinnerTail v-else size="20" color="white" />
+                        </button>
+                        <button type="button"
+                            class="inline-flex w-full justify-center rounded-md bg-blue-100 px-3 py-2 text-sm font-semibold text-blue-900 shadow-sm hover:bg-blue-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600"
+                            @click="kanban.openModal = false">
+                            <span>No</span>
+                        </button>
+                    </div>
+                </div>
+            </generic-modal>
+        </Teleport>
         
     </div>
 </template>
 
 <script setup>
-import { ref, onMounted, onUnmounted, nextTick } from 'vue'
+import { ref, onMounted, onUnmounted, nextTick, computed } from 'vue'
 import { useKanbanStore } from '../stores/kanban'
 import { DialogTitle, Listbox, ListboxButton, ListboxLabel, ListboxOption, ListboxOptions } from '@headlessui/vue'
-import { CheckIcon, ChevronUpDownIcon, TrashIcon } from '@heroicons/vue/20/solid'
+import { CheckIcon, ChevronUpDownIcon, TrashIcon } from '@heroicons/vue/24/solid'
+import { VueSpinnerTail } from 'vue3-spinners'
+import { useLoading } from 'vue-loading-overlay'
+import { toast } from 'vue3-toastify';
 import { sha256 } from 'js-sha256';
 
 
 const kanban = useKanbanStore()
 const selected = ref(null)
 const errors = ref(null)
+const taskPropsName = computed({
+    get() {
+        return kanban.creatingTask ? kanban.creatingTaskProps.name : kanban.updatingTaskProps.name;
+    },
+    set(val) {
+        kanban.creatingTask ? kanban.creatingTaskProps.name = val : kanban.updatingTaskProps.name = val;
+    }
+})
+const taskPropsUser = computed({
+    get() {
+        return kanban.creatingTask ? kanban.creatingTaskProps.user_id : kanban.updatingTaskProps.user_id;
+    },
+    set(val) {
+        kanban.creatingTask ? kanban.creatingTaskProps.user_id = val : kanban.updatingTaskProps.user_id = val;
+    }
+})
+const taskPropsPhase = computed({
+    get() {
+        return kanban.creatingTask ? kanban.creatingTaskProps.phase_id : kanban.updatingTaskProps.phase_id;
+    },
+    set(val) {
+        kanban.creatingTask ? kanban.creatingTaskProps.phase_id = val : kanban.updatingTaskProps.phase_id = val;
+    }
+})
+
+const $loading = useLoading()
 
 const getAvatar = function (user) {
     if (user.profile_picture_url !== null) {
@@ -142,7 +203,7 @@ const getAvatar = function (user) {
 
 const getError = function (field) {
     if (errors.value && errors.value[field]) {
-        return errors.value[field][0].message;
+        return errors.value[field].message;
     }
     return null;
 }
@@ -191,7 +252,19 @@ const mouseUpHandler = function () {
 
 const refreshTasks = async () => {
     try {
-        const response = await axios.get('/api/tasks');
+        let response;
+        if (kanban.pageLoaded) {
+            const loader = $loading.show({
+                backgroundColor: "#4B4B4B",
+                color: "#FFF",
+                opacity: 0.5
+            });
+            response = await axios.get('/api/tasks');
+            loader.hide();
+        } else {
+            response = await axios.get('/api/tasks');
+        }
+
         const originalTasks = response.data;
         kanban.phases = originalTasks.reduce((acc, cur) => {
             acc[cur.id] = cur;
@@ -224,26 +297,54 @@ const getSelf = async () => {
             kanban.creatingTaskProps.user_id = kanban.self.id;
         }
         if (kanban.self.profile_picture_url === null) {
-            kanban.self.profile_picture_url = getAvatar(kanban.self)
+            kanban.self.profile_picture_url = getAvatar(kanban.self);
         }
     } catch (error) {
         console.error('There was an error fetching the logged in user!', error);
     }
 }
 
-const addCard = async () => {
-    try {
-        const response = await axios.post('/api/tasks', kanban.creatingTaskProps);
-        kanban.creatingTask = false;
-        kanban.creatingTaskProps = {
-            name: null,
-            phase_id: null,
-            user_id: null
+const validateForm = () => {
+    if (taskPropsName.value === '') {
+        errors.value = {
+            name: {
+                message: 'This field is required'
+            }
         };
-        await refreshTasks();
-    } catch (error) {
-        if (error.response.status === 422) {
-            errors.value = error.response.data.errors;
+        return false;
+    } else {
+        errors.value && delete errors.value.name
+    }
+    return true;
+}
+
+const addCard = async () => {
+    if (validateForm()) {
+        try {
+            kanban.submittingTask = true
+            if (kanban.creatingTask) {
+                const response = await axios.post('/api/tasks', kanban.creatingTaskProps);
+            } else {
+                const response = await axios.put('/api/tasks', kanban.updatingTaskProps);
+            }
+
+            await refreshTasks();
+            toast.success(kanban.creatingTask ? 'Task created successfully!' : 'Task updated successfully!', {
+                autoClose: 3000,
+                position: toast.POSITION.TOP_RIGHT,
+                theme: 'colored'
+            })
+            resetTasks();
+        } catch (error) {
+            kanban.submittingTask = false
+            if (error.response.status === 422) {
+                errors.value = error.response.data.errors;
+            }
+            toast.error('Something went wrong!', {
+                autoClose: 3000,
+                position: toast.POSITION.TOP_RIGHT,
+                theme: 'colored'
+            })
         }
     }
 }
@@ -258,11 +359,59 @@ const deleteCard = async (id) => {
     }
 }
 
-onMounted(async () => {
+const confirmAction = async () => {
+    try {
+        kanban.confirmingAction = true;
+        if (kanban.markingPhase) {
+            const response = await axios.put('api/phases', { phase_id: kanban.phaseInAction });
+        } else {
+            const response = await axios.delete(`api/phases/${kanban.phaseInAction}`);
+        }
+        resetAction();
 
+        refreshTasks();
+    } catch (error) {
+        console.log('Something went wrong!');
+    }
+}
+
+const resetTasks = () => {
+    errors.value && delete errors.value.name;
+    kanban.submittingTask = false;
+    kanban.creatingTask = false;
+    kanban.creatingTaskProps = {
+        name: '',
+        phase_id: null,
+        user_id: null
+    };
+    kanban.updatingTask = false;
+    kanban.updatingTaskProps = {
+        name: '',
+        phase_id: null,
+        user_id: null
+    };
+}
+
+const resetAction = () => {
+    kanban.openModal = false;
+    kanban.markingPhase = false;
+    kanban.confirmingAction = false;
+    kanban.phaseInAction = null;
+}
+
+onMounted(async () => {
+    kanban.pageLoaded = false;
+    const loader = $loading.show({
+        backgroundColor: "#4B4B4B",
+        color: "#FFF",
+        opacity: 0.5
+    });
     await refreshTasks();
     await refreshUsers();
     await getSelf();
+
+    kanban.pageLoaded = true;
+    loader.hide();
 
     await nextTick();
 
@@ -282,7 +431,7 @@ onUnmounted(() => {
 })
 </script>
 
-<style scoped>/* For Webkit-based browsers (Chrome, Safari and Opera) */
+<style>/* For Webkit-based browsers (Chrome, Safari and Opera) */
 .scrollbar-hide::-webkit-scrollbar {
     display: none;
 }