+++ CHANGES.md
@@ -2,3 +2,17 @@
 
 ## 0.1
 - Initial release
+
+## 0.2
+# What's new?
+- Added a edit function to modify task
+- Added a card count after phase's name
+- Added a function mark one or more phases as being a 'completion'
+- Allowed the user to delete an entire column, and all cards in it
+- Integrated the "Tooltip" to icons
+- Integrated the "Toast" to notify user after action
+- Integrated the "Overlay-loading" to prevent unnecessary actions while fetching data via APIs
+- Integrated the "Loading button" to prevent multiple requests for same action
+- Added single-column vertical scroll to the board
+- Created the new page to see the statistics
+- Chagned the UI/UX for better experience